Update transformer code
